Added settings page, saved keys page
Fixed some typos, was messing around with trying to get everything stateful but I didn't have enough time for that.
Ran a linter which messed a bit with the position of comments on things that were too long.
I've temporarily removed the width resizer but I will add that back in later together with a manual selector.