solved some text overflow issues and set max tag name
Text overflow issues on the slot manager should be a lot better now.
Text overflow on saved cards still requires some work. Changed a row to a wrap to get it to handle overflows properly but I need to split the quick buttons to a new row now.
On the card edit page I added a validation for a max name length as supported by the chameleon ultra and added some localization translations for this.