Adjusted scrolling behavior in Saved cards
Now allows scrolling in saved cards and contains them within the Card item except for when it is fully scrolled to the top. I am not sure why it is doing this. I may wrap it in a ListView or SingleScrollableView container but those were giving some issues for now.