Added scaling for Android for reading cards
I added scaling for the reading cards page for android. Yes. This is based on the screen width. That page is so layered in unconstrained containers that this was legitimately the best option..... If you have a better suggestion feel free to adjust the layoutbuilder widget.
I also added a const in front of the complementary colors to improve the speed slightly.