Added height to slot_manager and setup checkmark scaling in read card
Do we want scaling like the saved cards on the slot manager page in addition to text overflow? Above a certain overflow it becomes really useless and scaling is preferred in my opinion.
For the read card section I have added a layout builder. It does nothing for now until the parent of the checkmarks gets constraints. This will need to be added at some point. The alternative to Transform is to use a FittedBox but that has the same issue. The parent isn't contrained. Alternatively, it is possible to grab the window width and the side bar width as a stopgap method.